+++ src/components/editpanel.component.js
@@ -1,5 +1,3 @@
-import linkH from '../static/linklist-type1.png'
-import linkV from '../static/linklist-type2.png'
 import prof1 from '../static/profile-type1.png'
 import prof2 from '../static/profile-type2.png'
 
@@ -32,6 +30,7 @@ import Link from "../router/link";
 import {colours, styles} from "../pages/profileDesigns/colour.util";
 import {IoIosList, IoMdAdd, IoMdCloudUpload} from "react-icons/io";
 import GenericPanel from "./panels/generic.panel.component";
+import LinkListPanel from "./panels/linklist.panel.component";
 
 const importAll = (r) => r.keys().map(r);
 const postFiles = importAll(require.context("../news/", true, /\.md$/))
@@ -57,12 +56,6 @@ export default class EditPanel extends React.Component
             displayName: "",
             userMessage: null,
             lastReloaded: Date.now(),
-            // Link list
-            selectedLinkListItem: null,
-            linkItemTitleField: "",
-            linkItemURLField: "",
-            linkItemSelectedImage: null,
-            linkItemMessage: null,
             // Spotify
             spotifyMessage: null,
             spotifyLink: "",
@@ -73,8 +66,7 @@ export default class EditPanel extends React.Component
 
         this.handleLinkFieldChange = this.handleLinkFieldChange.bind(this)
         this.handleDisplayNameChange = this.handleDisplayNameChange.bind(this)
-        this.handleLinkItemURLChange = this.handleLinkItemURLChange.bind(this)
-        this.handleLinkItemTitleChange = this.handleLinkItemTitleChange.bind(this)
+
         this.handleSpotifyLinkChange = this.handleSpotifyLinkChange.bind(this)
         this.handleYouTubeLinkChange = this.handleYouTubeLinkChange.bind(this)
     }
@@ -207,161 +199,6 @@ export default class EditPanel extends React.Component
         </div>
     }
 
-    addNewLinkItem = (component) =>
-    {
-        if (this.state.selectedLinkListItem !== null)
-        {
-            return this.displayLinkItemMessage({
-                type: 'error',
-                message: 'You must save your changes before adding another item!'
-            })
-        }
-        if (component.content.length >= 3)
-            return
-        const content = component.content
-        content.links.push({"url": "https://www.rar.vg", "icon": null, "title": "Lorem ipsum dolor sit amet"})
-        this.props.updateLocallyWithoutCancelling(content).then(r =>
-            this.selectLinkItem(component, content.links.length - 1))
-    }
-
-    checkURLValidity(url)
-    {
-        let url_;
-
-        try
-        {
-            url_ = new URL(url);
-        } catch (_)
-        {
-            return false;
-        }
-
-        return url_.protocol === "http:" || url_.protocol === "https:";
-    }
-
-    updateLinkItem = (component, link) =>
-    {
-        if (this.state.selectedLinkListItem !== null)
-        {
-            if (this.state.linkItemURLField === null)
-            {
-                return this.displayLinkItemMessage({type: 'error', message: 'Link field must not be empty!'})
-            }
-            if (this.checkURLValidity(this.state.linkItemURLField) === false)
-            {
-                return this.displayLinkItemMessage({type: 'error', message: 'Use the correct link format!'})
-            }
-            const newLink = {
-                url: this.state.linkItemURLField || link.url,
-                title: this.state.linkItemTitleField || link.title,
-                icon: link.icon
-            }
-            if (this.state.linkItemSelectedImage !== null)
-            {
-                return this.uploadLinkItemIcon().then(result =>
-                    {
-                        newLink.icon = config('HOST') + "/uploads/" + result
-                        const content = component.content
-                        content.links[this.state.selectedLinkListItem] = newLink
-                        this.props.updateLocallyWithoutCancelling(content).then(res =>
-                        {
-                            return this.setState({
-                                selectedLinkListItem: null,
-                                linkItemTitleField: null,
-                                linkItemURLField: null,
-                                linkItemSelectedImage: null
-                            })
-                        })
-                    }
-                )
-            }
-            const content = component.content
-            content.links[this.state.selectedLinkListItem] = newLink
-            this.props.updateLocallyWithoutCancelling(content).then(res =>
-            {
-                return this.setState({
-                    selectedLinkListItem: null,
-                    linkItemTitleField: null,
-                    linkItemURLField: null,
-                    linkItemSelectedImage: null
-                })
-            })
-        }
-    }
-
-    deleteLinkItem = (key, component) =>
-    {
-        const content = component.content
-        content.links.splice(key, 1)
-        this.props.updateLocallyWithoutCancelling(content).then(res =>
-        {
-            this.setState({
-                selectedLinkListItem: null,
-                linkItemTitleField: null,
-                linkItemURLField: null
-            })
-        })
-    }
-
-    onItemIconChange = (event) =>
-    {
-        if (event.target.files && event.target.files[0])
-        {
-            const allowedFiles = ['jpg', 'jpeg', 'png']
-            if (event.target.files && event.target.files[0])
-            {
-                if (!allowedFiles.includes(event.target.files[0].name.split('.').pop()))
-                    return this.displayLinkItemMessage({
-                        type: 'error',
-                        message: 'The selected file format is not allowed!'
-                    })
-                if (event.target.files[0].size / 1024 / 1024 > 1)
-                    return this.displayLinkItemMessage({type: 'error', message: 'The selected file is too large!'})
-                this.setState({linkItemSelectedImage: URL.createObjectURL(event.target.files[0])});
-            }
-        }
-    }
-
-    uploadLinkItemIcon = () =>
-    {
-        return new Promise(res =>
-        {
-            fetch(this.state.linkItemSelectedImage).then(r => r.blob()).then(blob =>
-            {
-                const result = new File([blob], "image.png", {type: 'application/png'})
-                upload(result, false).then(result =>
-                {
-                    if (result.success)
-                    {
-                        this.uploadingDialog.close()
-                        return res(result.content)
-                    }
-                })
-            })
-        })
-    }
-
-    selectLinkItem = (component, key) =>
-    {
-        if (this.state.selectedLinkListItem !== null)
-        {
-            return this.displayLinkItemMessage({
-                type: 'error',
-                message: 'You must save your changes before selecting another item!'
-            })
-        }
-        return this.setState({
-            selectedLinkListItem: key,
-            linkItemTitleField: component.content.links[key].title,
-            linkItemURLField: component.content.links[key].url
-        })
-    }
-
-    displayLinkItemMessage = (message) =>
-    {
-        this.setState({linkItemMessage: message})
-        setTimeout(() => this.setState({linkItemMessage: null}), 5000)
-    }
 
     displayPDFMessage = (message) =>
     {
@@ -374,7 +211,6 @@ export default class EditPanel extends React.Component
         this.setState({userMessage: message})
         setTimeout(() => this.setState({userMessage: null}), 5000)
     }
-
 
 
     displaySpotifyMessage = (message) =>
@@ -398,50 +234,6 @@ export default class EditPanel extends React.Component
         )
     }
 
-    linkEditItem = (link, key, selected, component) =>
-    {
-        if (selected)
-            return <div className="special-mock">
-                {this.drawMessage(this.state.linkItemMessage)}
-                <h2 className="s">Title:</h2>
-                <input className="input" onChange={this.handleLinkItemTitleChange}
-                       defaultValue={this.state.linkItemTitleField} type="text"
-                       placeholder="My awesome link"/>
-                <h2 className="s">Link:</h2>
-                <input className="input" required onChange={this.handleLinkItemURLChange} type="url"
-                       placeholder="https://yourwebsite.com" defaultValue={this.state.linkItemURLField}/>
-                <h2 className="s">Icon:</h2>
-                <div className={"button-center"}>
-                    {this.state.linkItemSelectedImage !== null || link.icon !== null ? (
-                        <img className={"icon"} src={this.state.linkItemSelectedImage || link.icon}
-                             alt={'Link icon'}/>) : <></>}
-                    <label htmlFor="link-icon-button" className="button-label">Upload icon</label>
-                    <input type={"file"} onChange={this.onItemIconChange} className="file-button"
-                           accept={".jpg,.png,.jpeg"}
-                           id="link-icon-button"/>
-                </div>
-                <div className="link-list-btn-container">
-                    <button className="button delete" onClick={() => this.deleteLinkItem(key, component)}>Delete
-                    </button>
-                    <button className="button" onClick={() => this.updateLinkItem(component, link)}>Done
-                    </button>
-                </div>
-            </div>
-        else return <div className={"inner-mock"}>
-            <div className={"hero"}>
-                {link.icon !== null ? (<img className={"link-icon"} src={link.icon} alt={'Link icon'}/>) :
-                    <div className={"bump"}></div>}
-            </div>
-            <div className={"link-content"}>
-                <span>{link.title || link.link}</span>
-                <div className={"link-list-btn-container"}>
-                    <button className={"icon-button"} onClick={() => this.selectLinkItem(component, key)}>
-                        <AiFillEdit/>
-                    </button>
-                </div>
-            </div>
-        </div>
-    }
 
     deselectItem = () =>
     {
@@ -543,15 +335,6 @@ export default class EditPanel extends React.Component
         this.setState({displayName: event.target.value})
     }
 
-    handleLinkItemTitleChange(event)
-    {
-        this.setState({linkItemTitleField: event.target.value})
-    }
-
-    handleLinkItemURLChange(event)
-    {
-        this.setState({linkItemURLField: event.target.value})
-    }
 
     handleSpotifyLinkChange(event)
     {
@@ -750,22 +533,23 @@ export default class EditPanel extends React.Component
                 </>
             case 'generic':
                 return <GenericPanel component={this.props.selectedComponent}
-                    drawMessage={this.drawMessage} deleteSelectedComponent={this.props.deleteSelectedComponent}
-                    cancel={this.cancel} saveLocally={this.saveLocally}
+                                     drawMessage={this.drawMessage}
+                                     deleteSelectedComponent={this.props.deleteSelectedComponent}
+                                     cancel={this.cancel} saveLocally={this.saveLocally}
                 />
             case 'sociallinks':
                 return <>
-                    <h3 className="m p-no-margin-top p-no-margin-bottom">Edit social links</h3>
-                    <div className="icon-list-div">{this.drawIcons()}</div>
-                    <h2 className="s p-no-margin-bottom p-no-margin-top title">Links:</h2>
-                    {this.props.user.sociallinks.map((link, key) => (
-                        <div>{this.socialLinkEditItem(link, key, this.state.selectedLink === key)}</div>))}
-                    <div className={"button-container"}>
-                        <button className="button delete-button"
-                                onClick={() => this.props.deleteSelectedComponent()}>Delete component
-                        </button>
-                        <button className="button" onClick={() => this.cancel()}>Done</button>
-                    </div>
+                <h3 className="m p-no-margin-top p-no-margin-bottom">Edit social links</h3>
+                <div className="icon-list-div">{this.drawIcons()}</div>
+                <h2 className="s p-no-margin-bottom p-no-margin-top title">Links:</h2>
+                {this.props.user.sociallinks.map((link, key) => (
+                    <div>{this.socialLinkEditItem(link, key, this.state.selectedLink === key)}</div>))}
+                <div className={"button-container"}>
+                    <button className="button delete-button"
+                            onClick={() => this.props.deleteSelectedComponent()}>Delete component
+                    </button>
+                    <button className="button" onClick={() => this.cancel()}>Done</button>
+                </div>
                 </>
             case 'pdf':
                 return <>
@@ -793,34 +577,12 @@ export default class EditPanel extends React.Component
                     </div>
                 </>
             case 'linklist':
-                return <>
-                    <dialog ref={ref => this.uploadingDialog = ref} className={"dashboard-modal"}>
-                        <span className={"m"}>Uploading...</span>
-                    </dialog>
-                    <h3 className="m p-no-margin-top p-no-margin-bottom">Edit link list</h3>
-                    {(component.content !== null) ? component.content.links.map((link, key) => this.linkEditItem(link, key, this.state.selectedLinkListItem === key, component)) : <></>}
-                    {(component.content !== null) ? component.content.links.length >= 5 ? <></> :
-                        <button className="inner-mock3" onClick={() => this.addNewLinkItem(component)}>
-                            <span className="mm p-no-margin-bottom p-no-margin-top">+</span>
-                        </button> : <></>}
-                    <p className="mm p-no-margin-top p-no-margin-bottom">Change list design</p>
-                    <div className='list-button-container'>
-                        <button className="button unraised link-img" type="button"
-                                onClick={() => this.setLinkListVertical(false)}>
-                            <img src={linkH} alt={'Horizontal'}/>
-                        </button>
-                        <button style={{marginLeft: "10%"}} className="button unraised link-img"
-                                onClick={() => this.setLinkListVertical(true)}>
-                            <img src={linkV} alt={'Vertical'}/>
-                        </button>
-                    </div>
-                    <div className={"button-container"}>
-                        <button className="button delete-button"
-                                onClick={() => this.props.deleteSelectedComponent()}>Delete component
-                        </button>
-                        <button className="button" onClick={() => this.cancel()}>Done</button>
-                    </div>
-                </>
+                return <LinkListPanel component={this.props.selectedComponent}
+                                      drawMessage={this.drawMessage}
+                                      deleteSelectedComponent={this.props.deleteSelectedComponent}
+                                      cancel={this.cancel} saveLocally={this.saveLocally}
+                                      updateLocallyWithoutCancelling={this.props.updateLocallyWithoutCancelling}
+                />
             case 'youtube':
                 return <>
                     <h3 className="m p-no-margin-top p-no-margin-bottom">Edit YouTube video</h3>
